requirements_prod.txt
-e ./src/harrastuspassi

django==2.2
django-filter==2.2.0
django-redis-sessions==0.5.0
djangorestframework==3.10.0
python-memcached==1.57
redis==2.10.6
newrelic==4.20.1.121
raven==6.10.0
whitenoise==4.1.3
gunicorn==19.9.0
django-mptt==0.10.0

# Must match Bergenia infra
psycopg2==2.7.4
Pillow==6.1.0

